~!~!~!~!~!~!~!~!~!~BIG CHANGES~!~!~!~!~!~!~!~!~!~
Total rework of the dialogue box and conversation load order. Moved dialogueBoxes to GameState and removed their place from SA/update to a function within GameState called advanceConversation which keeps track of the current state of the conversation (UI) and increments it. Definitely a fix for the drunk dad glitch, disagr, and other weird speech box shenanigans.